added (default) feature sha3_tests to enable custom keccak, fixed all tests
