+++ basic_bootloader/src/bootloader/block_header.rs
@@ -1,7 +1,6 @@
 use crate::bootloader::rlp;
 use arrayvec::ArrayVec;
-use crypto::sha3::Keccak256;
-use crypto::MiniDigest;
+use crypto::{MiniDigest, sha3::Keccak256};
 use ruint::aliases::{B160, U256};
 use zk_ee::utils::Bytes32;
