+++ basic_system/src/system_functions/keccak256.rs
@@ -44,8 +44,7 @@ fn keccak256_as_system_function_inner<D: ?Sized + Extend<u8>, R: Resources>(
     let native_cost = keccak256_native_cost::<R>(src.len());
     resources.charge(&R::from_ergs_and_native(ergs_cost, native_cost))?;
 
-    use crypto::sha3::*;
-    use crypto::MiniDigest;
+    use crypto::{MiniDigest, sha3::Keccak256};
     let mut hasher = Keccak256::new();
     hasher.update(src);
     let hash = hasher.finalize();
