+++ crypto/Cargo.toml
@@ -73,10 +73,12 @@ rand_core = { version = "0.6.4", features = ["getrandom"] }
 # forward = ["blake2"]
 forward = []
 proving = ["single_round_with_control", "bigint_ops", "secp256k1-static-context"]
-default = ["forward", "secp256k1-static-context"]
+# default = ["forward", "secp256k1-static-context"]
+default = ["forward", "secp256k1-static-context", "sha3_tests"] # this one has custom keccak precompile
 # bigint_ops = ["zeroize", "num-bigint", "num-traits", "num-integer"]
 bigint_ops = []
 single_round_with_control = ["blake2s_u32/blake2_with_compression"]
 blake2s_tests = []
 secp256k1-static-context = []
 alloc = []
+sha3_tests = []
