+++ evm_interpreter/src/utils.rs
@@ -18,8 +18,7 @@ pub fn bytereverse_u256(value: &mut U256) {
 }
 
 pub fn evm_bytecode_hash(bytecode: &[u8]) -> [u8; 32] {
-    use crypto::sha3::Keccak256;
-    use crypto::MiniDigest;
+    use crypto::{MiniDigest, sha3::Keccak256};
     let hash = Keccak256::digest(bytecode);
     let mut result = [0u8; 32];
     result.copy_from_slice(hash.as_slice());
