+++ tests/instances/erc20/markers.bench
@@ -10,8 +10,8 @@ Spent ergs for [keccak]: 10752
 
 Spent native for [keccak]: 20000
 
-Spent ergs for [process_tranSpent ergs for [run_prepared]: 5581312
-ocess_transaction]: 1959250
+Spent ergs for [process_transaction]: 17179904
+Spent native for [process_transaction]: 1959250
 Spent ergs for [ecrecover]: 768000
 
 Spent native for [ecrecover]: 522000
@@ -28,7 +28,10 @@ Spent ergs for [keccak]: 10752
 
 Spent native for [keccak]: 20000
 
-Spent ergs for [keccak]: 10Spent ergs for [run_prepared]: 23399424
-pent ergs for [process_transaction]: 13421056
+Spent ergs for [keccak]: 10752
+
+Spent native for [keccak]: 20000
+
+Spent ergs for [process_transaction]: 13421056
 Spent native for [process_transaction]: 1305784
 Spent ergs for [run_prepared]: 30600960
