+++ tests/instances/forge_tests/markers.bench
@@ -2,16 +2,14 @@ Spent ergs for [ecrecover]: 768000
 
 Spent native for [ecrecover]: 522000
 
-Spent ergs for [process_transaction]: 94453760
-Spent native for [process_transaction]: 3018672
-Spent ergs for [ecrecover]: 768000
+Spent ergs for [process_transaction]: 114272512
+Spent native for [process_transaction]: 1991065SSpent ergs for [ecrecover]: 768000
 
 Spent native for [ecrecover]: 522000
 
 Spent ergs for [prepare_for_call]: 0
 
 Spent native for [prepare_for_call]: 4500
-
 SSpent ergs for [prepare_for_call]: 25600
 
 Spent native for [prepare_for_call]: 4500
@@ -21,4 +19,5 @@ Spent ergs for [prepare_for_call]: 25600
 Spent native for [prepare_for_call]: 4500
 
 Spent ergs for [process_transaction]: 86038784
-Spent naSpent ergs for [run_prepared]: 160291328Spent ergs for [run_prepared]: 200311296
+Spent native for [process_transaction]: 2227286
+Spent ergs for [run_prepared]: 200311296
