+++ tests/instances/transactions/markers.bench
@@ -1,6 +1,6 @@
-Spent ergs for [ecrecover]: 76800Spent ergs for [run_prepared]: 0
-Spent ergs for [run_prepared]: 0
-Spent ergs for [run_prepared]: 0
+Spent ergs for [ecrecover]: 76800Spent ergs for [prepare_for_call]: 0
+
+Spent native for [prepare_foSpent ergs for [run_prepared]: 0
 keccak]: 10752
 
 Spent native for [keccak]: 20000
